requirements.txt
environs==9.5.0
aiogram~=2.25.1
vk_api==11.9.9
google-cloud-dialogflow==2.22.0
